confirm modal cleanup + linting
